updated install instructions

+++ dist.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-rm -rf dist
-rm tracker.zip
+if [ -e dist ]; then
+    rm -rf dist
+fi
+if [ -e tracker.zip ]; then
+    rm tracker.zip
+fi
 mkdir dist
 
 cp -r src/* dist
 cp README.markdown dist
 cp LICENSE dist
 cp doc/haxelib.xml dist
+cp bin/tracker.n dist/run.n
 
 cd dist
-mv tracker.n run.n
 rm test.n
 zip -r ../tracker.zip *
 cd ..
